0.3.8 RC5
+ replaced wplotgenerator with XPlotter v1.0 ->
https://github.com/Blagodarenko/Xplotter
+ individual pool input with port
+ new pools
+ implementation of dividends+ addon (Local Wallet)
+ Plotting expert mode: more than 32 threads
+ window state is saved

+++ src/Unit11.pas
@@ -104,12 +104,16 @@ end;
 
 procedure TForm11.Button1Click(Sender: TObject);
 begin
+if MessageDlg('Are you sure that you want to delete and resync the Blockchain?', mtConfirmation, [mbyes, mbcancel], 0) = mrYes then
+begin
 Killtask('javaw.exe');
 Sleep(600);
 DeleteDir('burst_db');
 Sleep(1000);
  WinExec('run_java_autodetect.bat', SW_HIDE);
 Label2.Visible:=true;
+Button2.Caption:='Close';
+end;
 end;
 
 procedure TForm11.Button2Click(Sender: TObject);
